Update player controller to not slide UP
When a steep slope was hit, player character would slide up.
This is due to applying force against steep slope even if not grounded.

Current solution is to disable movement inputs if going up even
when not grounded for .2 second.
This way we can still control character even in air, but when sliding
up, the control will be disabled to ground character.